+++ src/Ext/TEvent/Body.h
@@ -54,6 +54,7 @@ enum PhobosTriggerEvent
 	HouseDoesntOwnTechnoType = 602,
 	CellHasTechnoType = 604,
 	CellHasAnyTechnoTypeFromList = 605,
+	AttachedIsUnderAttachedEffect = 606,
 
 	_DummyMaximum,
 };
@@ -96,6 +97,8 @@ public:
 
 	static bool CellHasAnyTechnoTypeFromListTEvent(TEventClass* pThis, ObjectClass* pObject, HouseClass* pHouse);
 	static bool CellHasTechnoTypeTEvent(TEventClass* pThis, ObjectClass* pObject, HouseClass* pHouse);
+
+	static bool AttachedIsUnderAttachedEffectTEvent(TEventClass* pThis, ObjectClass* pObject);
 
 
 	class ExtContainer final : public Container<TEventExt>